// features/sdnr/odlux/odlux/apps/microwaveApp/src/linkCalculator/model/waveguide.ts
/**
 * ============LICENSE_START========================================================================
 * ONAP : ccsdk feature sdnr wt odlux
 * =================================================================================================
 * Copyright (C) 2021 highstreet technologies GmbH Intellectual Property. All rights reserved.
 * =================================================================================================
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License
 * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 * or implied. See the License for the specific language governing permissions and limitations under
 * the License.
 * ============LICENSE_END==========================================================================
 */
export type Waveguide = {
  modelId: number;
  id: number;
  modelName: string;
  type: {
    id: number;
    name: string;
    description: string;
  } | null;
  operationalParameters: WaveguideOperationalParameters | null;
  calculationParameters: WaveguideCalculationPrameters | null;
};
export type WaveguideLoss = {
  waveguideLossPerMeter: number;
  waveguideLossdB: number;
  waveguideName: string;
  waveguideLength: number;
  waveguideType: string;
  waveguideLossA: number;
  waveguideLossB: number;

};

export type WaveguideOperationalParameters = {
  band: string;
  type: string;
  shape: string;
  length: number;
  loss: string;

};

export type WaveguideCalculationPrameters = {
  waveguideLength: number; 
};